Refactored client_dir directory and units_factory. Fixed swapping of units in client_main_window
