+++ battle.py
@@ -10,7 +10,7 @@ from client_dir.settings import ANY_UNIT, CLOSEST_UNIT, \
 from battle_logging import logging
 from units_dir.models import CurrentDungeon, Player2Units, AllUnits
 from units_dir.units import main_db
-from units_dir.units_factory import Unit
+from units_dir.battle_unit import Unit
 
 PLAYER1_NAME = 'Erepb-89'
 PLAYER2_NAME = 'Computer'
