+++ battle_log.txt
@@ -1,3 +1,8 @@
-Одержимый наносит урон 29 воину Сквайр. Осталось ХП: 71
+Ходит: Титан
+Одержимый наносит урон 28 воину Титан. Осталось ХП: 194
 Ходит: Одержимый
+Адская гончая наносит урон 30 воину Архимаг. Осталось ХП: 35
+Адская гончая наносит урон 28 воину Титан. Осталось ХП: 222
+Адская гончая наносит урон 27 воину Ученик. Осталось ХП: 8
+Ходит: Адская гончая
 Новая битва
