+++ client_dir/ui/army.ui
