+++ client_dir/ui/army_d.ui
