+++ client_dir/ui/campaign.ui
