+++ client_dir/ui/capital_army.ui
