+++ client_dir/ui/capital_building.ui
