+++ client_dir/ui/capital_main.ui
