+++ client_dir/ui/choose_faction.ui
