+++ client_dir/ui/client_main.ui
