+++ client_dir/ui/hire_menu.ui
