+++ client_dir/ui/message.ui
