+++ client_dir/ui/unit_details.ui
