+++ client_dir/ui_functions.py
@@ -5,8 +5,9 @@ import os
 from PyQt5 import QtWidgets
 from PyQt5.QtGui import QMovie, QPixmap
 
-from client_dir.settings import UNIT_ICONS, PLUG, ICON, GIF_ANIMATIONS, \
-    HIRE_SCREEN, COMMON, UNIT_STAND, BIG, BATTLE_ANIM, UNIT_FACES, CURSOR
+from client_dir.settings import UNIT_ICONS, PLUG, ICON, \
+    GIF_ANIMATIONS, HIRE_SCREEN, COMMON, UNIT_STAND, \
+    BIG, BATTLE_ANIM, UNIT_FACES, CURSOR
 
 
 def get_image(folder: str, faction: str):
