+++ client_dir/windows/message_window.py
@@ -1,66 +1,56 @@
-"""Окно вопроса"""
+"""Окно сообщения"""
 import os
 
 from PyQt5 import QtCore
 from PyQt5.QtGui import QPixmap
 from PyQt5.QtWidgets import QMainWindow, QHBoxLayout
 
-from client_dir.question_form import Ui_Question
+from client_dir.forms.message_form import Ui_Message
 from client_dir.settings import CAPITAL_BUILDING
 
 
-class QuestionWindow(QMainWindow):
+class MessageWindow(QMainWindow):
     """
-    Класс - окно вопроса.
+    Класс - окно сообщения.
     Конфигурация окна создана в QTDesigner и загружается из
-    конвертированного файла question_form.py
+    конвертированного файла message_form.py
     """
 
-    def __init__(self, parent: any, text: str):
+    def __init__(self, parent_window: any, text: str):
         super().__init__()
         # основные переменные
-        self.parent = parent
+        self.parent_window = parent_window
         self.text = text
 
         self.InitUI()
 
     def InitUI(self):
         """Загружаем конфигурацию окна из дизайнера"""
-        self.ui = Ui_Question()
+        self.ui = Ui_Message()
         self.ui.setupUi(self)
 
         self.hbox = QHBoxLayout(self)
         self.update_bg()
 
-        self.ui.questionText.setText(self.text)
+        self.ui.messageText.setText(self.text)
 
-        self.ui.pushButtonYES.clicked.connect(
-            self.yes_action)
         self.ui.pushButtonNO.clicked.connect(
             self.no_action)
 
         self.show()
 
-    def yes_action(self):
-        """Согласиться"""
-        self.parent.question = True
-        self.parent.confirmation()
-        self.close()
-
     def no_action(self):
-        """Отказаться"""
-        self.parent.question = False
-        self.parent.confirmation()
+        """Закрыть"""
         self.close()
 
     def update_bg(self) -> None:
         """Обновление бэкграунда, заполнение картинкой"""
-        question_bg = self.ui.questionBG
-        question_bg.setPixmap(
+        message_bg = self.ui.messageBG
+        message_bg.setPixmap(
             QPixmap(
                 os.path.join(
                     CAPITAL_BUILDING,
                     "question.png")))
-        question_bg.setGeometry(QtCore.QRect(0, 0, 552, 461))
-        self.hbox.addWidget(question_bg)
+        message_bg.setGeometry(QtCore.QRect(0, 0, 552, 461))
+        self.hbox.addWidget(message_bg)
         self.setLayout(self.hbox)
